Fix a crash due to uneven KVO observer calls
If the application wasn't active when the FormViewController was
displayed, FieldRow would observe titleLabel.text twice (once in setup and once in the
NSNotification callback). It would only call removeObserver once in
deinit, which lead to a crash when the underlying label
was deallocated and KVO noticed the dangling observation.