Correctly handle nil values for `titleFieldPercentage`
If `titleFieldPercentage` is not set, the layout is not adjusted to make
up for potential accessory and/or image views. Instead a fixed constant
is used for the respective constraint. By moving this fixed constant to
the advanced layout calculation we not only solve the aforementioned
issue but also eliminate duplication of the the constant during
constraint setup.

Note that this could have also been achieved by simply declaring
`titleFieldPercentage` a non-optional variable and assigning a default
value of 0.3. Since this would represent both a code breaking as well as
somewhat of a semantic change the solution at hand was deemed preferable