Allow non-selectable rows to exist besides selectable rows in a selectable section
From time to time it is helpful to have non-selectable rows like switch rows besides selectable rows (e.g. as inline rows). While this basically works, it fails when the selected row changes. In this case, also the values of the non-selectable rows are reset. With a little check this can be fixed.