Update cell when tintColor changes
So the textColor can be updated if it uses the tintColor, which a lot of cells do. It appears that the tintColor is currently used before the cell is added to the window, so if the tintColor has been customised on the window the cells aren’t updated.

Calling row.updateCell() seems like the best thing to call, as that should encompass all code (built-in and custom) that uses the tintColor.